Shill Bidding
About this file
Abstract: It is scraped a large number of eBay auctions of a popular product. After preprocessing the auction data, to build the SB dataset. The goal is to share the labelled SB dataset with the researchers.

Attribute Information:

Record ID: Unique identifier of a record in the dataset.
Auction ID: Unique identifier of an auction.
Bidder ID: Unique identifier of a bidder.
Bidder Tendency: A shill bidder participates exclusively in auctions of few sellers rather than a diversified lot. This is a collusive act involving the fraudulent seller and an accomplice.
Bidding Ratio: A shill bidder participates more frequently to raise the auction price and attract higher bids from legitimate participants.
Successive Outbidding: A shill bidder successively outbids himself even though he is the current winner to increase the price gradually with small consecutive increments.
Last Bidding: A shill bidder becomes inactive at the last stage of the auction (more than 90\% of the auction duration) to avoid winning the auction.
Auction Bids: Auctions with SB activities tend to have a much higher number of bids than the average of bids in concurrent auctions.
Auction Starting Price: a shill bidder usually offers a small starting price to attract legitimate bidders into the auction.
Early Bidding: A shill bidder tends to bid pretty early in the auction (less than 25\% of the auction duration) to get the attention of auction users.
Winning Ratio: A shill bidder competes in many auctions but hardly wins any auctions.
Auction Duration: How long an auction lasted.
Class: 0 for normal behaviour bidding; 1 for otherwise.